src/requirements-frozen.txt
# Frozen requirements - exact versions for maximum stability
# This file locks all dependencies to the versions that are known to work.
# Use this for production deployments or when you want guaranteed compatibility.
#
# Install with: pip install -r requirements-frozen.txt

fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.50.0
openai-whisper==20250625
pydub==0.25.1
python-multipart==0.0.21
resemblyzer==0.1.4
audioop-lts; python_version >= "3.13"

# Core dependencies with pinned versions
numpy==1.26.4
torch==2.9.1
librosa==0.11.0
soundfile==0.13.1
scipy==1.15.3
scikit-learn==1.7.2
numba==0.63.1
llvmlite==0.46.0

# Audio processing
audioread==3.1.0
soxr==1.0.0
webrtcvad==2.0.10

# Whisper dependencies
tiktoken==0.12.0
regex==2026.1.15
tqdm==4.67.1

# Networking
requests==2.32.5
urllib3==2.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11

# Utilities
click==8.3.1
colorama==0.4.6
joblib==1.5.3
networkx==3.4.2
packaging==25.0
filelock==3.20.3
fsspec==2026.1.0
pooch==1.8.2
platformdirs==4.5.1
lazy_loader==0.4
decorator==5.2.1
msgpack==1.1.2
mpmath==1.3.0
sympy==1.14.0
threadpoolctl==3.6.0
more-itertools==10.8.0

# Type checking and annotations
typing_extensions==4.15.0
typing==3.7.4.3
typing-inspection==0.4.2
annotated-types==0.7.0
annotated-doc==0.0.4

# Server infrastructure
h11==0.16.0
anyio==4.12.1
exceptiongroup==1.3.1

# Build tools (if needed)
cffi==2.23
pycparser==2.23
setuptools==65.5.0

# Testing (optional - already in requirements-dev.txt)
# pytest==9.0.2
# pluggy==1.6.0
# iniconfig==2.3.0
# tomli==2.4.0

# Template engine (for UI)
Jinja2==3.1.6
MarkupSafe==3.0.3
Pygments==2.19.2
